# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Geoindicadores",
        "description": "Projeto onde armazena indicadores ambientais para análise",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/indicadores/{id}": {
            "get": {
                "tags": [
                    "Indicadores"
                ],
                "summary": "Pega a lista de Indicadores",
                "description": "Retorna a lista de Indicadores",
                "operationId": "de7380811dd6727c968a33b913dacf7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Um exemplo de resultado",
                                        "value": {
                                            "id": "integer",
                                            "nome": "string",
                                            "imagem": "string",
                                            "nota_tecnica": "string",
                                            "observacao": "string",
                                            "projeto_id": "integer",
                                            "projeto": "string",
                                            "fonte": {
                                                "id": "integer",
                                                "nome": "string",
                                                "descricao": "string"
                                            },
                                            "departamento": {
                                                "id": "integer",
                                                "sigla": "string",
                                                "nome": "string"
                                            },
                                            "periodicidade": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não Encontrado"
                    }
                }
            }
        },
...
    }
}